+++ courtlistener/__init__.py
@@ -1,10 +1,12 @@
 from courtlistener.alerts import DocketAlerts, SearchAlerts
 from courtlistener.citation_lookup import CitationLookup
 from courtlistener.client import CourtListener
+from courtlistener.exceptions import CourtListenerAPIError
 
 __all__ = [
     "CitationLookup",
     "CourtListener",
+    "CourtListenerAPIError",
     "DocketAlerts",
     "SearchAlerts",
 ]